moved TxError into its own error module
