+++ src/token/api/Cargo.toml
@@ -25,7 +25,7 @@ ic-helpers = { git = "https://github.com/infinity-swap/canister-sdk", package = 
 ic-storage = { git = "https://github.com/infinity-swap/canister-sdk", package = "ic-storage", tag = "v0.3.14" }
 ic-certified-map = "0.3"
 serde_cbor = "0.11"
-
+thiserror = "1.0"
 [target.'cfg(not(target_family = "wasm"))'.dependencies]
 async-std = {version = "1.10.0", features = ["attributes"]}
