+++ src/token/api/src/lib.rs
@@ -4,5 +4,6 @@ pub mod principal;
 pub mod state;
 pub mod types;
 
+pub mod error;
 #[cfg(test)]
 pub mod mock;
